+++ docs/internal/app/guides/home.component.tsx
@@ -39,6 +39,9 @@ export function Home() {
                 <p>Dig into the most common Roxie components</p>
                 <dl>
                   <dd><Link to="/components/dialog">Dialog</Link></dd>
+                  <dd><Link to="/components/dropdown">Dropdown</Link></dd>
+                  <dd><Link to="/components/tabpanel">TabPanel</Link></dd>
+                  <dd><Link to="/components/list">List</Link></dd>
                   <dd><Link to="/components/grid">Grid</Link></dd>
                   <dd><Link to="/components/chart">Chart</Link></dd>
                 </dl>