+++ packages/components/src/index.tsx
@@ -3,6 +3,9 @@ import { render } from 'react-dom';
 
 export * from './container.component';
 export * from './dialog.component';
+export * from './dropdown.component';
+export * from './tabpanel.component';
+export * from './list.component';
 export * from './grid.component';
 export * from './chart.component';